+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/CardHandler.kt
@@ -5,13 +5,13 @@ import ayds.lisboa.songinfo.moredetails.model.entities.Card
 import ayds.lisboa.songinfo.moredetails.model.entities.EmptyCard
 
 interface CardHandler {
-    fun navigateToLastFMActivity()
+   /* fun navigateToLastFMActivity()
     fun navigateToWikipediaActivity()
-    fun navigateToNYTActivity()
+    fun navigateToNYTActivity()*/
 }
 
-internal class CardHandlerImpl (
-    private val moreDetailsView: MoreDetailsView,
+/*internal class CardHandlerImpl (
+    /*private val moreDetailsView: MoreDetailsView,
     private val cards: List<Card>) : CardHandler {
     override fun navigateToLastFMActivity() {
         moreDetailsView.openCardActivity(getLastFMCard())
@@ -24,15 +24,5 @@ internal class CardHandlerImpl (
         moreDetailsView.openCardActivity(getNYTCard())
     }
 
-    private fun getLastFMCard() : Card {
-        return cards.firstOrNull{it.source == Source.LASTFM} ?: EmptyCard
-    }
 
-    private fun getWikipediaCard() : Card {
-        return cards.firstOrNull{it.source == Source.WIKIPEDIA} ?: EmptyCard
-    }
-
-    private fun getNYTCard() : Card {
-        return cards.firstOrNull{it.source == Source.NEWYORKTIMES} ?: EmptyCard
-    }
-}
+}*/
