+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/MoreDetailsUiEvent.kt
@@ -2,7 +2,5 @@ package ayds.lisboa.songinfo.moredetails.view
 
 sealed class MoreDetailsUiEvent {
     object Search: MoreDetailsUiEvent()
-    object OpenLastFM: MoreDetailsUiEvent()
-    object OpenWikipedia: MoreDetailsUiEvent()
-    object OpenNYT: MoreDetailsUiEvent()
+    object OpenSource: MoreDetailsUiEvent()
 }
