+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/view/MoreDetailsUiState.kt
@@ -1,9 +1,11 @@
 package ayds.lisboa.songinfo.moredetails.view
 
+import ayds.lisboa.songinfo.moredetails.model.entities.Card
+
 const val URL_IMAGE = "https://upload.wikimedia.org/wikipedia/commons/thumb/d/d4/Lastfm_logo.svg/320px-Lastfm_logo.svg.png"
 
 data class MoreDetailsUiState (
     val artistName: String = "",
     var actionsEnabled: MutableList<Boolean> = mutableListOf(false, false, false),
-
+    var cardActual: Card? = null
 )
